simplify KeeperTestSuite
